mods

+++ Python/ejercicio8.py
@@ -2,9 +2,9 @@ print("----------------------------------------------------")
 print("Ejercicio8: CALCULAR PERIMETRO Y SUPERFICIE DEL RECTANGULO")
 print("----------------------------------------------------")
 print("Ingrese la base: ")
-BASE = int(input())
+BASE = float(input())
 print("Ingrese la altura: ")
-ALTO = int(input())
+ALTO = float(input())
 SUP = BASE * ALTO
 PER = (BASE + ALTO)*2
 print("La superficie es de: " , SUP)
